Effects - remove more BigInt

+++ src/main/java/io/github/syst3ms/skriptparser/effects/EffEscape.java
@@ -9,7 +9,6 @@ import io.github.syst3ms.skriptparser.lang.control.Finishing;
 import io.github.syst3ms.skriptparser.lang.control.SelfReferencing;
 import io.github.syst3ms.skriptparser.parsing.ParseContext;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
 /**
@@ -17,10 +16,10 @@ import java.util.Optional;
  * Note that lines that have more indentation than this one are not considered in the amount.
  * Nevertheless, lines that have less indentation will be considered.
  *
+ * @author Mwexim
  * @name Escape
  * @pattern escape %integer% [(level[s]|line[s])]
  * @since ALPHA
- * @author Mwexim
  */
 public class EffEscape extends Effect {
     static {
@@ -30,12 +29,12 @@ public class EffEscape extends Effect {
         );
     }
 
-    private Expression<BigInteger> amount;
+    private Expression<Integer> amount;
 
     @SuppressWarnings("unchecked")
     @Override
     public boolean init(Expression<?>[] expressions, int matchedPattern, ParseContext parseContext) {
-        amount = (Expression<BigInteger>) expressions[0];
+        amount = (Expression<Integer>) expressions[0];
         return true;
     }
 
@@ -58,20 +57,21 @@ public class EffEscape extends Effect {
                 ((Finishing) current.get()).finish();
             }
             current = current.flatMap(val -> val instanceof SelfReferencing
-                    ? ((SelfReferencing) val).getActualNext()
-                    : val.getNext());
+                ? ((SelfReferencing) val).getActualNext()
+                : val.getNext());
 
             // Because otherwise SelfReferencing sections will first reference to themselves
             if (current.filter(val -> val instanceof Finishing).isEmpty())
                 am--;
         }
         return current.flatMap(val -> val instanceof SelfReferencing
-                ? ((SelfReferencing) val).getActualNext()
-                : val.getNext());
+            ? ((SelfReferencing) val).getActualNext()
+            : val.getNext());
     }
 
     @Override
     public String toString(TriggerContext ctx, boolean debug) {
         return "escape " + amount.toString(ctx, debug) + " lines";
     }
+
 }
